Renamed doxygen configfile to include a .dox extention for windows ease of use. This allows windows to map the .dox extension to doxywizard so that clicking on the Doxyfile.dox file will automatically bring up doxywizard. Note: the initial .dox extension mapping is manual as doxygen does not automatically install an association.

+++ glcd/build/distbuild/mkzip.bat
@@ -114,7 +114,7 @@ REM #
 
 set DOXYGEN=doxygen
 set DOXYGENDIR= %GLCDDISTDIR%\doc
-set DOXYGENCFG=Doxyfile
+set DOXYGENCFG=Doxyfile.dox
 
 REM #
 REM # Name of ZIP file
@@ -186,7 +186,7 @@ cd %PWD%
 
 echo Creating Zip file
 echo ======== Creating Zip file >> %LOGFILE%
-erase /F %GLCDZIPNAME%
+if exist %GLCDZIPNAME% erase /F %GLCDZIPNAME%
 echo About to create zip file from %GLCDDISTDIR%
 pause
 %ZIPCMD% %GLCDZIPNAME% %GLCDDISTDIR% >> %LOGFILE%
